Partitions removal logic extracted to a separate class
SpaceMaker was too big, now it's more focused in the process of
making space itself, not in the gory details.
